Make tests less fragile

+++ common/lib/xmodule/xmodule/modulestore/tests/django_utils.py
@@ -159,7 +159,7 @@ class ModuleStoreTestCase(TestCase):
             connection.drop_database(store.db.name)
             connection.close()
 
-        if store.contentstore.fs_files:
+        if hasattr(store, 'contentstore') and store.contentstore.fs_files:
             db = store.contentstore.fs_files.database
             db.connection.drop_database(db)
             db.connection.close()
